refactor data types
